添加了用户登入功能

+++ src/main/java/com/tz/warehouse/sys/entity/SysPermission.java
@@ -4,8 +4,9 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-import java.io.Serializable;
 import lombok.Data;
+
+import java.io.Serializable;
 
 /**
  * 
@@ -18,12 +19,12 @@ public class SysPermission implements Serializable {
      * 
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 
      */
-    private Integer pid;
+    private Long pid;
 
     /**
      * 权限类型[menu/permission]
